20240519 v5.11 more SQL kinds, no more mysql[sh] clients used
1. mysqlsh and mysql clients are no longer used for anything in imatest.pl
2. support for more SQL statement kinds
3. some refactoring
4. complete support for database reusal
5. better following changes in db structure during the test
6. ongoing improvements in log mining

+++ testright.sh
@@ -4,8 +4,8 @@ forhelp="$1"
 verbose="${2:-0}"
 seeds=""
 [ -z "$3" ] || seeds="--seed $3"
-echo "Usage: $0 [ports_rel [verbose_N [--seed N]]]"
+echo "Usage: $0 [ports_rel [verbose_N [seed_N]]]"
 echo "  so ports=$ports verbose=$verbose and seeds=$seeds"
-[ "$forhelp" = 'h' -o "$forhelp" = '-help' -o "$forhelp" = '-h' -o "$forhelp" = 'help' ] && exit 1
+[ -z "$forhelp" -o "$forhelp" = 'h' -o "$forhelp" = '-help' -o "$forhelp" = '-h' -o "$forhelp" = 'help' ] && exit 1
 sleep 3
 ./hakillms.sh "$ports" 9 wait;\rm -r /root/mysql-sandboxes/*/sandboxdata/error.log /tmp/*;./startms.sh "$ports" wait;./imatest.sh now --test imatest.yaml --verbose "$verbose" --nodry-run $seeds 2>&1|tee /tmp/test.out; ./stopms.sh "$ports" wait
